+++ src/components/Cards/LDO/LDOCard.tsx
@@ -34,8 +34,9 @@ interface LdoCardProperties {
     defaultSummary?: boolean;
 }
 
-const softStartItems = SoftStartValues.map(item => ({
-    label: `${item} mA`,
+const tempSoftStartValues = [10, 20, 35, 50];
+const softStartItems = SoftStartValues.map((item, index) => ({
+    label: `${tempSoftStartValues[index]} mA`,
     value: `${item}`,
 }));
 
@@ -139,21 +140,6 @@ export default ({
 
             {!summary && (
                 <>
-                    <Toggle
-                        label={
-                            <DocumentationTooltip
-                                card={card}
-                                item="SoftStartEnable"
-                            >
-                                Soft Start Enable
-                            </DocumentationTooltip>
-                        }
-                        isToggled={ldo.softStartEnabled}
-                        onToggle={value =>
-                            npmDevice.setLdoSoftStartEnabled(index, value)
-                        }
-                        disabled={disabled || ldo.mode === 'LDO'}
-                    />
                     <Dropdown
                         label={
                             <DocumentationTooltip